refactor: type 파일 옮김
