+++ src/types/main.ts
@@ -31,23 +31,3 @@ export type TextSlicePrams = {
   limitLength: number;
   sliceLength: number;
 };
-
-export type WordDetail = {
-  id: string;
-  name: string;
-  description: string;
-  diacritic: string[];
-  pronunciation: string[];
-  wrongPronunciations: string[];
-  // \n으로 구분해 사용
-  exampleSentence: string;
-
-  // NOTE: metadata
-  createdAt: Date;
-  updatedAt: Date;
-};
-
-export type DefaultRes<TData> = {
-  status: number;
-  data: TData;
-};
